Run Prettier on HTML, CSS, JS, JSON, and MD files
This commit adds Prettier to the project and uses it to format HTML,
CSS, JavaScript, JSON, and Markdown. A "format" script is also added to
format from the command line. Most editors can format on save, but I
still find this to be useful.

I don't agree with Prettier on everything. I really wish it used
trailing commas and single quotes. But I think consistency with other
projects is helpful and beginning down the road of discussing different
configuration options just seems like a waste of time.

If haven't had much luck with Svelte plugins, so this commit doesn't
format any Svelte files. We can revisit Svelte formatting down the road.